Finished knightMoves() function
As far as I know, knightMoves is finished. I had
the misunderstanding that printing out the steps
to BFS was the solution but instead printed out
every possible move. BFS is going to get all
possible moves until we reach the end point,
but the trick here is that we need to track
how we got to the endpoint from previous moves,
so I made a vertex.js file that created vertex
objects that contain the value of the vertex,
and the previous move it came from (parentVertex).
So, we loop through ALL possible moves, and every
time we calculate the new moves, we assign that
the next move's parentVertex is the givenVertex
because all these new moves are originating from
that position. We then enqueue these moves onto
the queue for traversal while NOT queuing any
entries that have already been visited
(a vertex has been visited if we are calulating
new moves for it), so I add the givenVertex to
the visited list. So, as long as the move hasn't
already been visited before OR out of bounds,
the BFS algorithm will visit all possible moves
until it reaches the end position, which then we
can check the end position's parentVertex property
to track where it came from. Still need to test
it out fully, but works from position [0,0] to
[3,3].